chore: Move viz gallery components into better structure
