+++ src/components/pages/homepage/visualization-gallery/components/chart-data-link.module.scss
