+++ src/components/pages/homepage/visualization-gallery/components/sidebar.js
