+++ src/components/pages/homepage/visualization-gallery/index.js
@@ -3,12 +3,12 @@ import React from 'react'
 import { graphql, useStaticQuery } from 'gatsby'
 import { Tabs, TabList, Tab, TabPanels, TabPanel } from '@reach/tabs'
 import Container from '~components/common/container'
-import USMap from './gallery-items/us-map'
-import ChartNationalHospitalization from './gallery-items/charts/national-hospitalization'
-import ChartNationalCases from './gallery-items/charts/national-cases'
-import GalleryChartNationalDeaths from './gallery-items/charts/national-deaths'
-import GalleryChartNationalTests from './gallery-items/charts/national-tests'
-import NationalCharts from './gallery-items/national-chart'
+import USMap from './items/us-map'
+import ChartNationalHospitalization from './items/single-charts/national-hospitalization'
+import ChartNationalCases from './items/single-charts/national-cases'
+import GalleryChartNationalDeaths from './items/single-charts/national-deaths'
+import GalleryChartNationalTests from './items/single-charts/national-tests'
+import NationalCharts from './items/national-chart'
 import galleryStyle from './visualization-gallery.module.scss'
 
 const components = {
