+++ src/components/pages/homepage/visualization-gallery/items/national-chart.js
@@ -2,9 +2,9 @@ import React from 'react'
 import { useStaticQuery, graphql } from 'gatsby'
 import { DateTime } from 'luxon'
 import { Row, Col } from '~components/common/grid'
-import { NationalTotals } from './sidebar'
-import Title from './title'
-import ChartDataLink from './chart-data-link'
+import { NationalTotals } from '../components/sidebar'
+import Title from '../components/title'
+import ChartDataLink from '../components/chart-data-link'
 import colors from '~scss/colors.module.scss'
 import BarChart from '~components/charts/bar-chart'
 import nationalChartStyle from './national-charts.module.scss'
