+++ src/components/pages/homepage/visualization-gallery/items/national-charts.module.scss
