+++ src/components/pages/homepage/visualization-gallery/items/single-charts/chart.module.scss
