+++ src/components/pages/homepage/visualization-gallery/items/single-charts/national-hospitalization.js
