+++ src/components/pages/homepage/visualization-gallery/items/us-map/grid.module.scss
