+++ src/components/pages/homepage/visualization-gallery/items/us-map/sidebar.module.scss
