+++ src/components/pages/homepage/visualization-gallery/visualization-gallery.module.scss
